setup-environment-internal: fix whiptail and dialog command usage
Commit a6f29ed was a bit more too aggressive.. and adding quotes when calling
whiptail or dialog was wrong. These tools expect parameters to be passed by
tupple, and adding the quotes does not work. Since this commit, sourcing the
script with MACHINE and/or DISTRO not set would result in bad things such as:

    MACHINE = Box options:
    --msgbox <text> <height> <width>
    --yesno  <text> <height> <width>
    --infobox <text> <height> <width>
    --inputbox <text> <height> <width> [init]
    --passwordbox <text> <height> <width> [init]
    --textbox <file> <height> <width>
    --menu <text> <height> <width> <listheight> [tag item] ...
    --checklist <text> <height> <width> <listheight> [tag item status]...
    --radiolist <text> <height> <width> <listheight> [tag item status]...
    --gauge <text> <height> <width> <percent>
Options: (depend on box-option)
         --clear

Also make sure that MACHINETABLE and DISTROTABLE are unset before they are
used, if the script failed once it's possible to have them partially set in the
env..

Change-Id: Ib470731e5c3fb87c4405303517ba314576d8a555
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>