local.conf: do not enable PRSERV by default
Using PRSERV is a good idea *if* we are building an official distro and its
associated feed. It requires to save/restrore the PRSERV state to make sure we
don't break the versioning. It is not really a good idea to enable it by default
for everyone, since it can cause package numbers to go backwards.

Change-Id: I6cb10a6fb885d45b3b1b1e5bc193f52c9285239f
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>