setup-environment: update DISTRO logic, in case there is no distro conf file
When re-using setup-environment in a different configuration such as
OpenEmbedded Core with no specific distro (e.g. no conf/distro/*.conf files),
then DISTROLAYERS can be empty and the setup script start to fall apart...

This change will check for such a special case, and default to setting
DISTRO=nodistro, which is exactly what we need to do.

This has been tested with a minimal manifest file with only oe-core and
meta-qcom.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>