setup: Control failure if SRCREV_kernel is undefined
Containing the test for an undefined variable helps when
running bash with -e. In this case, when SRCREV_kernel is not
defined (as when building with linux-dummy) leads to an
unhandled error in "[ -n $2 ]" as only the successful case was
actually executed (i.e. there is no || in the command).

If'ing this test is a way of dealign with this problem.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>

(Cherry-picked from commit a4f498656e8c94080c1bc0d6e09c54c7ed1cf420)