.github: make local_manifest optional
Add support to build a manifest without a local_manifest, to allow a
user to make a vanilla build. We need to add option input params for
the repo and revision in the local_manifest.

While adding input params, also add a parameter for the target branch,
so let users pick the branch.

Rename the template, since it can support regular builds, not just
premerge.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>